// apps/admin/src/lib/api/generated/models/jobLogEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.1
 */
import type { JobLogEntryFields } from './jobLogEntryFields';
import type { JobLogEntryLevel } from './jobLogEntryLevel';

export interface JobLogEntry {
  fields: JobLogEntryFields;
  id: string;
  level: JobLogEntryLevel;
  message: string;
  timestamp: string;
}